moved close button and click to base class
